demo/tgbot-python-docker/src/requirements.txt
aiogram==2.19
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autopep8==1.5.7
Babel==2.9.1
beautifulsoup4==4.10.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.8
click==8.0.1
docutils==0.18.1
frozenlist==1.3.0
idna==3.3
image==1.5.33
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
multidict==6.0.2
netifaces==0.11.0
Pillow==8.4.0
plyer==2.0.0
psycopg2-binary==2.9.1
py3dbp==1.1.2
pycodestyle==2.7.0
Pygments==2.10.0
pyjnius==1.4.1
pyserial==3.5
python-dateutil==2.8.2
pytz==2021.3
requests==2.26.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.1
yarl==1.7.2
